[usage] Track last compelted ledger job time
